requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.16
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
h5py>=3.8.0
tensorboard>=2.13.0
